automation_orchestrator/requirements_extra.txt
fastapi>=0.110
uvicorn>=0.23
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31
openai>=1.0.0
tiktoken>=0.7.0
